% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp.R
\name{dsp.data.frame}
\alias{dsp.data.frame}
\title{dsp}
\usage{
\method{dsp}{data.frame}(data, species, distance)
}
\arguments{
\item{data}{a data frame}

\item{species}{data frame column specifying the species of the observation}

\item{distance}{distance from random sampling point to nearest individual of a species}
}
\value{
input data frame with Dsp (density of species) column added on
}
\description{
Density of species (Dsp) function
}
\examples{
df1 <- data.frame(sp = c("a", "b", "b", "a", "b", "a"),
                  dis = c(3, 6, 4, 7, 3, 8))
dsp(df1, sp, dis)
}
